[EST-005] - Implementação exceptionHandler

+++ src/main/java/com/pitang/testeTecnico/repository/UsuarioRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {}
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {}
